feat(viewer): TopBar de-clutter + DATA sidebar panel (v2 Phase 4)
Per spec §6 + decision D4, the TopBar is now informational chrome
only — title + tier indicator + location chip + EARTHDATE + search.
UPLOAD CLOUD, SCAN LOCAL, and DELETE actions move into a new DATA
panel triggered from a dedicated ACTIONS-group sidebar item.

Changes
- TopBar.tsx: stripped all data-action props (onCloudUpload,
  onRescan, onDeleteUnload, deleteCounts, hasCloudData, etc.).
  Added tierIndicator slot, locationLabel chip, EARTHDATE chip
  (auto-formats today's date as YYYY.MM.DD when no override
  given). RescanStatus / UploadStatus types still exported here
  since the host viewer wires them through to DataPanel.
- DataPanel.tsx (new): slide-in left-edge sheet with CLOUD /
  LOCAL / DELETE sections. Hosts the same state-aware Upload /
  Scan buttons that previously lived in the TopBar (label
  transitions and progress copy moved verbatim) plus the
  NuclearReset dropdown. Esc + scrim-click close.
- Sidebar.tsx: added optional onOpenDataPanel + dataPanelOpen
  props. New ACTIONS group renders the DATA item below
  BROWSE / INSIGHTS in the vertical variant; horizontal variant
  appends a DAT pill after the four mode pills.
- ChatArchViewer.tsx: hoisted dataPanelOpen state; wired the
  panel + the new TopBar slots; removed the LOG chip from the
  header (the ActivityLogPanel's own always-visible edge tab
  remains the entry point [INFERRED — spec §6 reads "only" but
  D4 only enumerated UPLOAD / SCAN / DELETE]).
- styles.css: added rules for the new TopBar chips
  (.lcars-top-bar__location, .lcars-top-bar__earthdate,
  .lcars-top-bar__tier-slot), the DATA item accent
  (.lcars-sidebar__group--actions, .lcars-sidebar__item--data),
  and the DataPanel sheet (scrim + dialog + section layout).

Tests
- Sidebar.test.tsx: added a DATA-panel-trigger describe block
  covering ACTIONS-group ordering, click handler invocation,
  aria-pressed reflection, and the horizontal-variant DAT pill.
- ChatArchViewer.test.tsx: the upload-precedence test now opens
  the DATA panel before driving the hidden file input, since the
  input no longer lives in the TopBar.

Validation suite (all green): pnpm -r typecheck, lint, test,
build, plus apps/standalone astro check. Visual snapshots at
mobile / tablet / desktop confirm zero source-btn elements
remain in .lcars-top-bar and the DATA panel renders the three
sections cleanly (see _planning/snapshot-phase4.mjs).
